chore: modularize config files
