+++ src/config/benefits.js
@@ -1,0 +1,23 @@
+export const benefits = [
+  {
+    image: './images/benefits/benefits-icon-hexagon.svg',
+    title: 'Top-Notch Technology',
+    text: "Once you transform your indoor logistics, we'll keep it up-to-date.",
+  },
+  {
+    image: './images/benefits/benefits-icon-cloud.svg',
+    title: 'Up-To-Date Software',
+    text: 'You always get immediate software updates including new features and algorithms in background.',
+  },
+  {
+    image: './images/benefits/benefits-icon-four-corners.svg',
+    title: 'Seamless Fleet Upgrade',
+    text:
+      'No need to stop the production to upgrade, extend, or replace the fleet. No hassle needed for the new units to get to work.',
+  },
+  {
+    image: './images/benefits/benefits-icon-numbers.svg',
+    title: 'Incompatible Robots Team-Up',
+    text: 'Robots produced by tens of different vendors can collaborate in the same workspace.',
+  },
+];
