+++ src/config/cards-with-images.js
@@ -1,0 +1,26 @@
+export const cardsWithImages = [
+  {
+    image: './images/workplace-images/man-on-pallet-jack.png',
+    title: 'Labor-intensive dolly movements',
+    text:
+      '**Challenges:** It is more and more **difficult to hire workers**, because they prefer to work in delivery services for the same money.',
+  },
+  {
+    image: './images/workplace-images/man-on-forklift.png',
+    title: 'Forklifts, tuggers managed by man',
+    text:
+      '**Challenges:** It is more and more **difficult to hire forklift drivers**  because they prefer to work in delivery for the same money.',
+  },
+  {
+    image: './images/workplace-images/follow-the-line-robots.png',
+    title: 'Follow-the-line robots',
+    text:
+      '**Challenges:** It is time-consuming and expensive to replace old robots by state-of-the art, but produced by another vendor.',
+  },
+  {
+    image: './images/workplace-images/autonomous-robots.png',
+    title: 'Autonomous mobile robots ',
+    text:
+      '**Challenges**: Limited upgrade options within the same vendor robot line only and too expensive to use another vendor robots.',
+  },
+];
