+++ src/config/products.js
@@ -1,0 +1,24 @@
+export const products = [
+  {
+    image: './public/images/products-1-OTTO.svg',
+    title: 'OTTO 100 (pending)',
+    data: {
+      'Max payload: ': '220 lbs (100kg)',
+      'Size: ': '740 x 550 x 310 mm',
+      'Max speed: ': '4.3 mph',
+      'Running time: ': '4 hrs or 9 mi',
+      'Charging time: ': '1.2hrs (0-80%: 0.75 hrs)',
+    },
+  },
+  // {
+  //   "image": "./public/images/products-2-placeholder.svg",
+  //   "title": "Placeholder",
+  //   "data": {
+  //     "Max payload: ": "220 lbs (100kg)",
+  //     "Size: ": "740 x 550 x 310 mm",
+  //     "Max speed: ": "4.3 mph",
+  //     "Running time: ": "4 hrs or 9 mi",
+  //     "Charging time: ": "1.2hrs (0-80%: 0.75 hrs)"
+  //   },
+  // }
+];
