--- src/utils/config.js
@@ -1,75 +1,0 @@
-export const benefits = [
-  {
-    image: './images/benefits/benefits-icon-hexagon.svg',
-    title: 'Top-Notch Technology',
-    text: "Once you transform your indoor logistics, we'll keep it up-to-date.",
-  },
-  {
-    image: './images/benefits/benefits-icon-cloud.svg',
-    title: 'Up-To-Date Software',
-    text: 'You always get immediate software updates including new features and algorithms in background.',
-  },
-  {
-    image: './images/benefits/benefits-icon-four-corners.svg',
-    title: 'Seamless Fleet Upgrade',
-    text:
-      'No need to stop the production to upgrade, extend, or replace the fleet. No hassle needed for the new units to get to work.',
-  },
-  {
-    image: './images/benefits/benefits-icon-numbers.svg',
-    title: 'Incompatible Robots Team-Up',
-    text: 'Robots produced by tens of different vendors can collaborate in the same workspace.',
-  },
-];
-
-export const products = [
-  {
-    image: './public/images/products-1-OTTO.svg',
-    title: 'OTTO 100 (pending)',
-    data: {
-      'Max payload: ': '220 lbs (100kg)',
-      'Size: ': '740 x 550 x 310 mm',
-      'Max speed: ': '4.3 mph',
-      'Running time: ': '4 hrs or 9 mi',
-      'Charging time: ': '1.2hrs (0-80%: 0.75 hrs)',
-    },
-  },
-  // {
-  //   "image": "./public/images/products-2-placeholder.svg",
-  //   "title": "Placeholder",
-  //   "data": {
-  //     "Max payload: ": "220 lbs (100kg)",
-  //     "Size: ": "740 x 550 x 310 mm",
-  //     "Max speed: ": "4.3 mph",
-  //     "Running time: ": "4 hrs or 9 mi",
-  //     "Charging time: ": "1.2hrs (0-80%: 0.75 hrs)"
-  //   },
-  // }
-];
-
-export const cardsWithImages = [
-  {
-    image: './images/workplace-images/man-on-pallet-jack.png',
-    title: 'Labor-intensive dolly movements',
-    text:
-      '**Challenges:** It is more and more **difficult to hire workers**, because they prefer to work in delivery services for the same money.',
-  },
-  {
-    image: './images/workplace-images/man-on-forklift.png',
-    title: 'Forklifts, tuggers managed by man',
-    text:
-      '**Challenges:** It is more and more **difficult to hire forklift drivers**  because they prefer to work in delivery for the same money.',
-  },
-  {
-    image: './images/workplace-images/follow-the-line-robots.png',
-    title: 'Follow-the-line robots',
-    text:
-      '**Challenges:** It is time-consuming and expensive to replace old robots by state-of-the art, but produced by another vendor.',
-  },
-  {
-    image: './images/workplace-images/autonomous-robots.png',
-    title: 'Autonomous mobile robots ',
-    text:
-      '**Challenges**: Limited upgrade options within the same vendor robot line only and too expensive to use another vendor robots.',
-  },
-];
